Переписать логику работы истории.
